[lldb-dap] Split the launch IO redirection tests.
Tests in the same python file share the same build directory.
although they are the same binary. The logs are different.
