+++ lldb/test/API/tools/lldb-dap/launch/io/DAP_launch_io.py
@@ -11,17 +11,8 @@ consoles, environments, stdin and cli arguments.
 """
 
 from abc import abstractmethod
-from typing import IO
 import lldbdap_testcase
 from tempfile import NamedTemporaryFile
-
-from lldbsuite.test.decorators import (
-    skip,
-    skipIfAsan,
-    skipIfBuildType,
-    skipIfRemote,
-    skipIfWindows,
-)
 
 
 class DAP_launchIO(lldbdap_testcase.DAPTestCaseBase):
@@ -210,152 +201,3 @@ class DAP_launchIO(lldbdap_testcase.DAPTestCaseBase):
         It requires subclasses to implement the specific mechanism for obtaining the stderr stream.
         """
         raise RuntimeError(f"NotImplemented for {self}")
-
-
-@skipIfWindows
-class TestDAP_launchInternalConsole(DAP_launchIO):
-    console = "internalConsole"
-    __debuggee_stdout = None
-
-    # all redirection
-    def test_all_redirection(self):
-        self.all_redirection(console=self.console)
-
-    def test_all_redirection_with_args(self):
-        self.all_redirection(console=self.console, with_args=True)
-
-    # stdin
-    def test_stdin_redirection(self):
-        self.stdin_redirection(console=self.console)
-
-    def test_stdin_redirection_with_args(self):
-        self.stdin_redirection(console=self.console, with_args=True)
-
-    # stdout
-    def test_stdout_redirection(self):
-        self.stdout_redirection(console=self.console)
-
-    def test_stdout_redirection_with_env(self):
-        self.stdout_redirection(console=self.console, with_env=True)
-
-    # stderr
-    def test_stderr_redirection(self):
-        self.stderr_redirection(console=self.console)
-
-    def test_stderr_redirection_with_env(self):
-        self.stderr_redirection(console=self.console, with_env=True)
-
-    def _get_debuggee_stdout(self) -> str:
-        # self.get_stdout is not idempotent.
-        if self.__debuggee_stdout is None:
-            self.__debuggee_stdout = self.get_stdout()
-        return self.__debuggee_stdout
-
-    def _get_debuggee_stderr(self) -> str:
-        # NOTE: In internalConsole stderr writes to stdout.
-        return self._get_debuggee_stdout()
-
-
-@skipIfRemote
-@skipIfAsan
-@skipIfBuildType(["debug"])
-@skipIfWindows
-class TestDAP_launchIntegratedTerminal(DAP_launchIO):
-    console = "integratedTerminal"
-
-    # all redirection
-    def test_all_redirection(self):
-        self.all_redirection(console=self.console)
-
-    def test_all_redirection_with_args(self):
-        self.all_redirection(console=self.console, with_args=True)
-
-    # stdin
-    def test_stdin_redirection(self):
-        self.stdin_redirection(console=self.console)
-
-    def test_stdin_redirection_with_args(self):
-        self.stdin_redirection(console=self.console, with_args=True)
-
-    # stdout
-    def test_stdout_redirection(self):
-        self.stdout_redirection(console=self.console)
-
-    def test_stdout_redirection_with_env(self):
-        self.stdout_redirection(console=self.console, with_env=True)
-
-    # stderr
-    def test_stderr_redirection(self):
-        self.stderr_redirection(console=self.console)
-
-    def test_stderr_redirection_with_env(self):
-        self.stderr_redirection(console=self.console, with_env=True)
-
-    def _get_debuggee_stdout(self) -> str:
-        self.assertIsNotNone(
-            self.dap_server.reverse_process, "Expected a debuggee process."
-        )
-        proc_stdout: IO = self.dap_server.reverse_process.stdout
-        self.assertIsNotNone(proc_stdout)
-        return proc_stdout.read().decode()
-
-    def _get_debuggee_stderr(self) -> str:
-        self.assertIsNotNone(
-            self.dap_server.reverse_process, "Expected a debuggee process."
-        )
-        proc_stderr = self.dap_server.reverse_process.stderr
-        self.assertIsNotNone(proc_stderr)
-        return proc_stderr.read().decode()
-
-
-@skip  # NOTE: Currently there is no difference between internal and externalTerminal.
-@skipIfRemote
-@skipIfAsan
-@skipIfBuildType(["debug"])
-@skipIfWindows
-class TestDAP_launchExternalTerminal(DAP_launchIO):
-    console = "externalTerminal"
-
-    # all redirection
-    def test_all_redirection(self):
-        self.all_redirection(console=self.console)
-
-    def test_all_redirection_with_args(self):
-        self.all_redirection(console=self.console, with_args=True)
-
-    # stdin
-    def test_stdin_redirection(self):
-        self.stdin_redirection(console=self.console)
-
-    def test_stdin_redirection_with_args(self):
-        self.stdin_redirection(console=self.console, with_args=True)
-
-    # stdout
-    def test_stdout_redirection(self):
-        self.stdout_redirection(console=self.console)
-
-    def test_stdout_redirection_with_env(self):
-        self.stdout_redirection(console=self.console, with_env=True)
-
-    # stderr
-    def test_stderr_redirection(self):
-        self.stderr_redirection(console=self.console)
-
-    def test_stderr_redirection_with_env(self):
-        self.stderr_redirection(console=self.console, with_env=True)
-
-    def _get_debuggee_stdout(self) -> str:
-        self.assertIsNotNone(
-            self.dap_server.reverse_process, "Expected a debuggee process."
-        )
-        proc_stdout: IO = self.dap_server.reverse_process.stdout
-        self.assertIsNotNone(proc_stdout)
-        return proc_stdout.read().decode()
-
-    def _get_debuggee_stderr(self) -> str:
-        self.assertIsNotNone(
-            self.dap_server.reverse_process, "Expected a debuggee process."
-        )
-        proc_stderr = self.dap_server.reverse_process.stderr
-        self.assertIsNotNone(proc_stderr)
-        return proc_stderr.read().decode()
